Splits state structs into new state.h/c files, adds modeling documention to sipnet.c
